Work out the function name with `complete -p`
Not all of the completion scripts create a completion function named
`_$command`, but not all of them (I tested with the list at [0]). Since
this module is intended to be used with any version of the
bash-completion scripts, we can't just keep a list or mapping, as it
could get out of date. Instead, the first time we run a completer,
source the script and parse the output of `complete -p $command` to find
out the name of the completion function it's created.

[0]: https://github.com/aca/elvish-bash-completion/blob/b6c8024a2a0d7105ecd300121d5d666d37563b4b/README.md#usage